Bug fixed PostConfirm class

+++ PostAds/Config/Confirm/PostConfirm.cs
@@ -1,11 +1,12 @@
 ﻿using Motorcycle.HTTP;
+using NLog;
 using OpenPop.Pop3;
 
 namespace Motorcycle.Config.Confirm
 {
-    static class PostConfirm
+    internal static class PostConfirm
     {
-        public static void GetMails(string hostname, int port, bool useSsl, string username, string password)
+        public static void ConfirmAdv(string hostname, int port, bool useSsl, string username, string password)
         {
             // The client disconnects from the server when being disposed
             using (var client = new Pop3Client())
@@ -50,13 +51,18 @@ namespace Motorcycle.Config.Confirm
 
                     var url = text.Substring(start, stop - start);
                     var req = Request.GETRequest(url);
-                    var resp = Response.GetResponse(req);
-                    resp.
+                    //var resp = Response.GetResponse(req);
 
+                    var Log = LogManager.GetCurrentClassLogger();
 
                     var respString = Response.GetResponseString(req);
 
-                    //TODO -> LOG
+                    if (respString.Contains("after_confirm=false"))
+                    {
+                        Log.Warn(string.Format("Confirmation of {0} failed", username));
+                        continue;
+                    }
+                    Log.Info(string.Format("Confirmation of {0} success", username));
 
                     client.DeleteMessage(i);
                 }
